backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
python-multipart==0.0.9
qdrant-client==1.10.1
openai==1.40.6
tiktoken==0.7.0
pypdf==4.3.1
python-docx==1.1.2
SQLAlchemy==2.0.32
alembic==1.13.2
rapidfuzz==3.9.6
